+++ Runtime/Models/PlayerDataModels.cs
@@ -1,17 +1,9 @@
 using System;
 using System.Collections.Generic;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 
 namespace Flock.Models
 {
-    public enum PlayerTemplateTag
-    {
-        gameplay,
-        currency,
-        achievement
-    }
-
     public class PlayerTemplateSchema
     {
         [JsonProperty("id")]
@@ -24,19 +16,13 @@ namespace Flock.Models
         public string GameVersionId { get; set; }
 
         [JsonProperty("schema")]
-        public Dictionary<string, object> Schema { get; set; }
+        public List<TypedSchema> Schema { get; set; }
 
         [JsonProperty("data")]
-        public Dictionary<string, object> Data { get; set; }
+        public List<DataField> Data { get; set; }
 
         [JsonProperty("tag")]
         public string Tag { get; set; }
-
-        public T GetDataAs<T>()
-        {
-            if (Data == null) return default;
-            return JObject.FromObject(Data).ToObject<T>();
-        }
     }
 
     public class PlayerData
@@ -54,30 +40,12 @@ namespace Flock.Models
         public string PlayerId { get; set; }
 
         [JsonProperty("data")]
-        public Dictionary<string, object> Data { get; set; }
+        public List<DataField> Data { get; set; }
 
         [JsonProperty("created_at")]
         public DateTime CreatedAt { get; set; }
 
         [JsonProperty("updated_at")]
         public DateTime UpdatedAt { get; set; }
-    }
-
-    internal class PlayerDataRequest
-    {
-        [JsonProperty("game_id")]
-        public string GameId { get; set; }
-
-        [JsonProperty("player_id")]
-        public string PlayerId { get; set; }
-
-        [JsonProperty("data")]
-        public Dictionary<string, object> Data { get; set; }
-    }
-
-    internal class UpdatePlayerDataRequest
-    {
-        [JsonProperty("data")]
-        public Dictionary<string, object> Data { get; set; }
     }
 }